refactor: move interp_2d to autoarray.operators; surface as aa.interp_2d
Relocates interp_2d from the new autoarray/numerics/ subpackage into the
existing operators/ subpackage. operators/ already houses similar
grid-data operations (Convolver, contour, transformer, over_sampling),
so a dedicated numerics/ subpackage for a single function was
over-engineering. Top-level surface is now aa.interp_2d (mirroring how
aa.Convolver is exposed from operators/convolver.py).

Issue PyAutoArray#306.
