--- autoarray/numerics/__init__.py
@@ -1,1 +1,0 @@
-from autoarray.numerics.interp_2d import interp_2d
